# CDRIVER-3645 Close operation connections to removed servers