--- a/src/libmongoc/doc/mongoc_client_pool_try_pop.rst
+++ b/src/libmongoc/doc/mongoc_client_pool_try_pop.rst
@@ -14,6 +14,10 @@ Synopsis
 
 This function is identical to :symbol:`mongoc_client_pool_pop()` except it will return ``NULL`` instead of blocking for a client to become available.
 
+.. note::
+
+   Return a checked out :symbol:`mongoc_client_t` to the pool with :symbol:`mongoc_client_pool_push` quickly to encourage reuse of clients among threads.
+
 Parameters
 ----------