--- a/src/libmongoc/src/mongoc/mongoc-client-pool.c
+++ b/src/libmongoc/src/mongoc/mongoc-client-pool.c
@@ -52,6 +52,8 @@ struct _mongoc_client_pool_t {
    bool error_api_set;
    mongoc_server_api_t *api;
    bool client_initialized;
+   // `last_known_serverids` is a sorted array of uint32_t.
+   mongoc_array_t last_known_serverids;
 };
 
 
@@ -146,6 +148,7 @@ mongoc_client_pool_new_with_error (const mongoc_uri_t *uri, bson_error_t *error)
    }
 
    pool = (mongoc_client_pool_t *) bson_malloc0 (sizeof *pool);
+   _mongoc_array_init (&pool->last_known_serverids, sizeof (uint32_t));
    bson_mutex_init (&pool->mutex);
    mongoc_cond_init (&pool->cond);
    _mongoc_queue_init (&pool->queue);
@@ -228,6 +231,8 @@ mongoc_client_pool_destroy (mongoc_client_pool_t *pool)
 #ifdef MONGOC_ENABLE_SSL
    _mongoc_ssl_opts_cleanup (&pool->ssl_opts, true);
 #endif
+
+   _mongoc_array_destroy (&pool->last_known_serverids);
 
    bson_free (pool);
 
@@ -357,6 +362,76 @@ mongoc_client_pool_try_pop (mongoc_client_pool_t *pool)
    RETURN (client);
 }
 
+typedef struct {
+   mongoc_array_t *server_ids;
+   mongoc_cluster_t *cluster;
+} prune_ctx;
+
+static int
+server_id_cmp (const void *a_, const void *b_)
+{
+   uint32_t *a = (uint32_t *) a_;
+   uint32_t *b = (uint32_t *) b_;
+
+   if (*a == *b) {
+      return 0;
+   }
+
+   return a < b ? -1 : 1;
+}
+
+static bool
+maybe_prune (void *item, void *ctx_)
+{
+   mongoc_cluster_node_t *cn = (mongoc_cluster_node_t *) item;
+   prune_ctx *ctx = (prune_ctx *) ctx_;
+   uint32_t server_id = cn->handshake_sd->id;
+
+   if (!bsearch (&server_id, ctx->server_ids->data, ctx->server_ids->len, sizeof (uint32_t), server_id_cmp)) {
+      mongoc_cluster_disconnect_node (ctx->cluster, server_id);
+   }
+   return true;
+}
+
+// `prune_client` closes connections from `client` to servers not contained in `known_server_ids`.
+static void
+prune_client (mongoc_client_t *client, mongoc_array_t *known_server_ids)
+{
+   BSON_ASSERT_PARAM (client);
+   BSON_ASSERT_PARAM (known_server_ids);
+
+   mongoc_cluster_t *cluster = &client->cluster;
+
+   bool needs_prune = false;
+   // Do a fast initial check to see if a prune is needed.
+   size_t idx1 = 0;
+   size_t idx2 = 0;
+   for (; idx1 < cluster->nodes->items_len; idx1++) {
+      // Compare both sorted lists in order. `cluster->nodes` may be a smaller list if not all servers were used.
+      mongoc_set_item_t *cn = &cluster->nodes->items[idx1];
+      bool found = false;
+      for (; idx2 < known_server_ids->len; idx2++) {
+         uint32_t last_known = _mongoc_array_index (known_server_ids, uint32_t, idx2);
+         if (cn->id == last_known) {
+            found = true;
+            break;
+         }
+      }
+      if (!found) {
+         // A server in the cluster is not in the last known server ids. Prune it.
+         needs_prune = true;
+         break;
+      }
+   }
+
+   if (!needs_prune) {
+      return;
+   }
+
+   prune_ctx ctx = {.cluster = cluster, .server_ids = known_server_ids};
+   mongoc_set_for_each (cluster->nodes, maybe_prune, &ctx);
+}
+
 
 void
 mongoc_client_pool_push (mongoc_client_pool_t *pool, mongoc_client_t *client)
@@ -370,6 +445,52 @@ mongoc_client_pool_push (mongoc_client_pool_t *pool, mongoc_client_t *client)
    mongoc_cluster_reset_sockettimeoutms (&client->cluster);
 
    bson_mutex_lock (&pool->mutex);
+   // Check if `last_known_server_ids` needs update.
+   bool serverids_have_changed = false;
+   {
+      mc_shared_tpld td = mc_tpld_take_ref (pool->topology);
+      const mongoc_set_t *servers = mc_tpld_servers_const (td.ptr);
+
+      if (pool->last_known_serverids.len != servers->items_len) {
+         serverids_have_changed = true;
+      } else {
+         // Check if any server IDs have changed.
+         for (size_t i = 0; i < servers->items_len; i++) {
+            // Compare in order. Both lists are sorted.
+            uint32_t last_known = _mongoc_array_index (&pool->last_known_serverids, uint32_t, i);
+            uint32_t actual = servers->items[i].id;
+            if (last_known != actual) {
+               serverids_have_changed = true;
+               break;
+            }
+         }
+      }
+
+      if (serverids_have_changed) {
+         // Store new set of server IDs.
+         _mongoc_array_destroy (&pool->last_known_serverids);
+         _mongoc_array_init (&pool->last_known_serverids, sizeof (uint32_t));
+         for (size_t i = 0; i < servers->items_len; i++) {
+            _mongoc_array_append_val (&pool->last_known_serverids, servers->items[i].id);
+         }
+      }
+      mc_tpld_drop_ref (&td);
+   }
+
+   // Check if pooled clients need to be pruned.
+   if (serverids_have_changed) {
+      // The set of last known server IDs has changed. Prune all clients in pool.
+      mongoc_queue_item_t *ptr = pool->queue.head;
+      while (ptr != NULL) {
+         prune_client ((mongoc_client_t *) ptr->data, &pool->last_known_serverids);
+         ptr = ptr->next;
+      }
+   }
+
+   // Always prune incoming client. The topology may have changed while client was checked out.
+   prune_client (client, &pool->last_known_serverids);
+
+   // Push client back into pool.
    _mongoc_queue_push_head (&pool->queue, client);
 
    if (pool->min_pool_size && _mongoc_queue_get_length (&pool->queue) > pool->min_pool_size) {