--- a/src/libmongoc/tests/TestSuite.c
+++ b/src/libmongoc/tests/TestSuite.c
@@ -1245,3 +1245,35 @@ test_bulkwriteexception_str (const mongoc_bulkwriteexception_t *bwe)
                    tmp_json (mongoc_bulkwriteexception_writeconcernerrors (bwe)),
                    tmp_json (mongoc_bulkwriteexception_errorreply (bwe)));
 }
+
+int32_t
+get_current_connection_count (const char *host_and_port)
+{
+   char *uri_str = bson_strdup_printf ("mongodb://%s\n", host_and_port);
+   char *uri_str_with_auth = test_framework_add_user_password_from_env (uri_str);
+   mongoc_client_t *client = mongoc_client_new (uri_str_with_auth);
+   test_framework_set_ssl_opts (client);
+   bson_t *cmd = BCON_NEW ("serverStatus", BCON_INT32 (1));
+   bson_t reply;
+   bson_error_t error;
+   bool ok = mongoc_client_command_simple (client, "admin", cmd, NULL, &reply, &error);
+   if (!ok) {
+      printf ("serverStatus failed: %s\n", error.message);
+      abort ();
+   }
+   int32_t conns;
+   // Get `connections.current` from the reply.
+   {
+      bson_iter_t iter;
+      BSON_ASSERT (bson_iter_init_find (&iter, &reply, "connections"));
+      BSON_ASSERT (bson_iter_recurse (&iter, &iter));
+      BSON_ASSERT (bson_iter_find (&iter, "current"));
+      conns = bson_iter_int32 (&iter);
+   }
+   bson_destroy (&reply);
+   bson_destroy (cmd);
+   mongoc_client_destroy (client);
+   bson_free (uri_str_with_auth);
+   bson_free (uri_str);
+   return conns;
+}