# Summary

Close operation connections to removed servers.

Verified with this patch build: https://spruce.mongodb.com/version/665489593f686b000739adf0

# Background & Motivation

Motivated by an issue reported in [HELP-59749](https://jira.mongodb.org/browse/HELP-59749). Tests observed an unexpected increasing connection count. This may have been triggered by (misconfigured?) DNS records resulting in servers being frequently added/removed from the topology during SRV polling.

## Rejected alternative: callback

I considered adding a callback to `_mongoc_topology_description_remove_server` to more immediately signal to close connections. This was rejected to reduce risk of an AB-BA deadlock.

`_mongoc_topology_description_remove_server` expects the caller to hold `mongoc_topology_t::tpld_modification_mtx`. Modifying pooled client requires holding `mongoc_client_pool_t::mutex`.

There is an existing pattern of locking `mongoc_client_pool_t::mutex` then `mongoc_topology_t::tpld_modification_mtx` (via `mongoc_client_pool_try_pop` => `_start_scanner_if_needed` => `_mongoc_topology_background_monitoring_start` => `mc_tpld_modify_begin`).

Instead, this PR proposes lazily pruning stale connections on `mongoc_client_pool_push`. I expect multi-threaded applications to pop/push clients frequently to encourage reuse of clients among threads. A note is added to documentation to make this recommendation explicit.

## Performance Impact

[A benchmark](https://spruce.mongodb.com/version/6653e2ad7f76a90007f5b338) was run to compare throughput of repeated use of the pool through pop+command+push with pruning:


| Test                                   |   Baseline (mb/sec) | With simple pruning   | With conditional pruning   |
|----------------------------------------|------------|-----------------------|----------------------------|
| Parallel/Pool/ShareClients/Threads:1   |     251995 | 236063 (-6.32%)       | 248131 (-1.53%)            |
| Parallel/Pool/ShareClients/Threads:10  |    1364224 | 873991 (-35.93%)      | 1337326 (-1.97%)           |
| Parallel/Pool/ShareClients/Threads:100 |    2017530 | 1050534 (-47.93%)     | 2020829 (0.16%)            |
| Parallel/Pool/ShareClients/Threads:150 |    1540427 | 1571373 (2.01%)       | 1552121 (0.76%)            |

An initial solution ("With simple pruning") iterated over all pooled clients to prune on each call to `mongoc_client_pool_push`. Due to the performance impact, checks were added to only prune if topology description change was detected ("With conditional pruning"). This motivated storing the `last_known_serverids` in the `mongoc_client_pool_t` to quickly determine if the set of server IDs changes. I expect the set of server IDs changing is a "rare" event (e.g. HELP-59749 notes an extreme case where SRV records unexpectedly change. Even then, I expect the set of server IDs would change about once per minute).